drm/msm: add idle hint for connectors
RSC already contains logic to power down during a defined idle time out.
Use the same logic to hint the connector that display is going idle and
potentially go to a lower power configuration. This also aligns well
with so that RSC reconfiguration is not needed while running in RSC CMD.

Bug: 129355653
Change-Id: I6ac987e4fec7224d7b8956944b17fcb4fff60aa3
Signed-off-by: Adrian Salido <salidoa@google.com>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>
Signed-off-by: CloudedQuartz <ravenklawasd@gmail.com>